Workaround Fix for #2
Changed some stuff around as well.

+++ src/main/java/com/yurisuika/blossom/mixin/block/LeavesBlockMixin.java
@@ -19,7 +19,7 @@ public class LeavesBlockMixin {
         if (state.isIn(BlockTags.LOGS)) {
             info.setReturnValue(0);
         } else if (state.getBlock() instanceof LeavesBlock || state.getBlock() instanceof FloweringLeavesBlock) {
-            info.setReturnValue((Integer)state.get(DISTANCE));
+            info.setReturnValue(state.get(DISTANCE));
         } else {
             info.setReturnValue(7);
         }
